requirements.txt
flask==2.2.2
slackclient==2.9.4
python-dotenv==0.21.0
slackeventsapi
flask-sqlalchemy==3.0.2
psycopg2==2.9.4
pytest==7.1.3
pytest-mock==3.10.0
pylint==2.15.3
sphinx==5.2.3
sphinx_rtd_theme==1.0.0
coverage==6.5.0
pytest-cov==4.0.0
pre-commit==2.20.0
flake8==5.0.4
black==22.8.0
slack-sdk
mock
sqlalchemy
pyee==11.0.0
Werkzeug==2.2.2
